separate github client into new package and add method to call CreateIssueComment API
